requirements.txt
absl-py>=0.9.0
flask>=1.1.2
numpy>=1.18.1
pandas>=1.0.4
protobuf>=3.12.0
protoc-wheel-0>=3.12.0
psycopg2-binary>=2.8.5
pygithub>=1.51
python-dateutil>=1.10.0
jinja2>=2.0.0
xlrd>=1.0.0
